Audit: 9 logic fixes, +105 tests, strict warnings clean
Comprehensive audit + fixes across the RIPscrip parser, renderer, and
font/icon modules.  ctest 3/3 suites pass (118 individual tests, up
from 13).  Build clean with -Wall -Wextra -Werror=implicit-function-
declaration on GCC 15.2.

Logic bugs (L1-L9):
- LINE_CONT now preserves \! \| \ escape pairs so unescape_text
  resolves them in text params (previously '|' triggered early
  dispatch, splitting them).
- Unified scale_y formula across ripscrip.c and ripscrip2.c so
  Level 0 'v' and Level 2 'P' map the same EGA Y to the same card
  pixel.
- Drop redundant outline pass in pie/oval-pie when filling.
- Fix BGI fill style → card pattern table (BGI 2 LINE was mapped to
  50% checker, remapped to horizontal); expose rip_bgi_fill_to_card
  so ripscrip2.c shares one source of truth.
- ripscrip2_init's port_alpha mirror now matches rip_port_t init.
- rip_init_first contract documented; uses base==NULL as init
  signal instead of trusting potentially-garbage size field.
- $ABORT$ and similar text-variable state resets now persist past
  the FSM's |/CR-LF dispatch handlers.

Mechanics (M2/M5/M6/M9/M11/M12/M14/M15):
- eval_if_expr checks 2-char ops (!=, >=, <=) before 1-char so 5>=5
  evaluates correctly.
- RIP_SCROLL strip-clear clamps to source rect.
- draw_thick_line switched to brush-stamp Bresenham for uniform
  perpendicular thickness at all angles.
- BGI italic does real per-stroke X shear inside render_char.
- bgi_font_t.num_chars promoted from u8 to u16.
- Patterned fills now use back_color for OFF bits per BGI/RIP
  semantics; 'k' propagates to draw layer; 'E' clears to back_color.
- '@' (RIP_TEXT_XY) now variable-expands consistent with 'T' via
  shared rip_render_text helper.

Syntax / cleanup (S1-S11):
- Extract clear_levels and rip_render_text helpers.
- Dedupe bgi_fill_to_card across TUs; merge O/V command handlers.
- Correct FSM state-count comments (14 states, not 13).
- Remove dead gpu_psram_alloc extern.
- Enable -Wall -Wextra -Werror=implicit-function-declaration.
- tests/CMakeLists.txt early-returns when BUILD_TESTING is OFF.
- rip_icn.c missing <stddef.h> for size_t.

Tests:
- New coverage: FSM transitions, variable expansion, Level 1 / 2
  commands, host callbacks, drawing primitives, mouse hit-test,
  font rendering edge cases.
- 5 compat fixtures with FNV-1a frame-hash lockdown (was 2).
- test_compat fixture path now via generated header for portability
  across MSYS/POSIX gcc invocations.

+++ include/bgi_font.h
@@ -15,8 +15,8 @@
  *   LITT = Small
  *   SCRI = Script
  *
- * Copyright (c) 2026 Brad Hawthorne
- * Licensed under GPL-3.0
+ * Copyright (c) 2026 SimVU (Brad Hawthorne)
+ * Licensed under the MIT License. See LICENSE.
  */
 
 #pragma once
@@ -41,13 +41,13 @@
 typedef struct {
     const uint8_t *data;     /* Raw .CHR binary data (after header) */
     int            data_size;
-    uint8_t        num_chars; /* Number of characters in font */
+    uint16_t       num_chars; /* Number of characters in font (on-disk u16 LE) */
     uint8_t        first_char;/* ASCII code of first character */
     int16_t        top;       /* Distance from origin to top */
     int16_t        baseline;  /* Distance from origin to baseline */
     int16_t        bottom;    /* Distance from origin to bottom */
     const uint8_t *widths;    /* Character width table */
-    const uint16_t *offsets;  /* Stroke data offset table */
+    const uint8_t *offsets;   /* LE stroke offset table (2 bytes per char) */
     const uint8_t *strokes;   /* Stroke data */
 } bgi_font_t;
